fix bugs with obsidian and refactoring
