+++ package.json
@@ -18,6 +18,7 @@
     "@types/lodash": "^4.14.176",
     "@types/node": "^14.17.21",
     "@types/parse5": "^6.0.1",
+    "@types/react-scroll": "^1.8.3",
     "@types/showdown": "^1.9.4"
   },
   "dependencies": {
