+++ src/components/ColumnDepth.tsx
@@ -7,6 +7,7 @@ import { Card } from './Card';
 export const ListColumnsDepths: FunctionComponent<IListColumnsDepths_Props> = ({ cards }) => {
   const $divColumn = useRef<HTMLDivElement>(null);
   const isSelectedCard = _.some(cards, { isSelected: true });
+  const columnDepth = _.map(cards, 'depth')[0];
 
   useEffect(() => {
     if ($divColumn && isSelectedCard) {
@@ -15,7 +16,7 @@ export const ListColumnsDepths: FunctionComponent<IListColumnsDepths_Props> = ({
   }, [$divColumn, isSelectedCard]);
 
   return (
-    <div className="column" ref={$divColumn}>
+    <div className="column" ref={$divColumn} id={`depth-${columnDepth}`}>
       <div className="empty-place"></div>
       <div className="group">
         {cards.map((card) => {
