+++ src/redux/reducers-case-logic/rootReducer/delete-card.case.ts
