+++ src/redux/reducers-case-logic/rootReducer/index.ts
@@ -1,3 +1,3 @@
-export * from './addCard.case.';
-export * from './changeCard.case';
-export * from './deleteCard.case';
+export * from './add-card.case.';
+export * from './change-card.case';
+export * from './delete-card.case';
