+++ styles.css
@@ -19,10 +19,12 @@
   padding: 10px 20px;
   font-weight: 600;
   overflow-x: auto;
+  scroll-snap-type: none;
 }
 
 .column {
   scroll-snap-type: y mandatory;
+  scroll-snap-align: center;
   overflow-x: hidden;
   overflow-y: scroll;
   scrollbar-width: none;
